+++ CryptocurrencyExchange/Models/FutureDto.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public string Symbol { get; set; } = string.Empty;
-        public decimal Margin { get; set; }
-        public decimal EntryPrice { get; set; }
+        public double Margin { get; set; }
+        public double EntryPrice { get; set; }
         public int Leverage { get; set; }
         public PositionType Position { get; set; }
